Refactor to group functions togeether
