+++ src/generate/generateAppRoutes/index.ts
@@ -1,0 +1,2 @@
+export { default } from './generateAppRoutes';
+export * from './generateAppRoutes';
