--- src/generate/processAppConfig/index.ts
@@ -1,2 +1,0 @@
-export { default } from './processAppConfig';
-export * from './processAppConfig';
